requirements.txt
requests>=2.28.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0

# Development/Testing dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
responses>=0.20.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
pre-commit>=3.3.0
coverage>=7.0.0